pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ckanext-harvest-basket"
version = "1.3.11"
description = "A bunch of custom harvesters for ckanext-harvest"
authors = [
    {name = "DataShades", email = "datashades@linkdigital.com.au"},
    {name = "Sergey Motornyuk", email = "sergey.motornyuk@linkdigital.com.au"},
    {name = "Oleksandr Cherniavskyi", email = "mutantsan@gmail.com"},
]
maintainers = [
    {name = "DataShades", email = "datashades@linkdigital.com.au"},
]
license = {text = "AGPL"}
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
keywords = ["CKAN"]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest-ckan",
    "mkdocs>=1.6.1,<1.7",
    "mkdocs-material>=9.5.42,<10",
    "mkdocs-include-markdown-plugin>=6.2.2,<7",
    "mkdocstrings-python>=1.11.1,<2",
    "mkdocs-exclude>=1.0.2,<2"
]
csw = [
    "OWSLib>=0.31.0,<1"
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
Homepage = "https://github.com/DataShades/ckanext-harvest-basket"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::sqlalchemy.exc.SADeprecationWarning",
    "ignore::sqlalchemy.exc.SAWarning",
    "ignore::DeprecationWarning",
]

[project.entry-points]
"babel.extractors" = {ckan = "ckan.lib.extract:extract_ckan"}

[project.entry-points."ckan.plugins"]
harvest_basket="ckanext.harvest_basket.plugin:HarvestBasketPlugin"
dkan_harvester="ckanext.harvest_basket.harvesters:DKANHarvester"
junar_harvester="ckanext.harvest_basket.harvesters:JunarHarvester"
socrata_harvester="ckanext.harvest_basket.harvesters:SocrataHarvester"
arcgis_harvester="ckanext.harvest_basket.harvesters:ArcGISHarvester"
custom_ckan_harvester="ckanext.harvest_basket.harvesters:CustomCKANHarvester"
ods_harvester="ckanext.harvest_basket.harvesters:ODSHarvester"
basket_dcat_json_harvester="ckanext.harvest_basket.harvesters:BasketDcatJsonHarvester"
basket_csw_harvester="ckanext.harvest_basket.harvesters:BasketCswHarvester"
basket_csiro_harvester="ckanext.harvest_basket.harvesters:CsiroHarvester"

[tool.setuptools.packages]
find = {}
